+++ src/core/transport.js
@@ -1,8 +1,9 @@
 class Transport{
-  constructor({from, to, args = []}){
+  constructor({from, to, args = [], ...options}){
     this.from = from;
     this.to = to;
     this.args = args;
+    this.options = options
   }
 
   transformArgsToPromptQuestion(){