+++ src/plugins/entity/attribute.js
@@ -1,7 +1,6 @@
 import fs from 'fs';
 import chalk from 'chalk';
 import EntityManager from './entityManager';
-import ConfigManager from '../../core/configManager';
 
 class Attribute{
     constructor(configPlugin, baseDir, render, options = {}){
@@ -11,7 +10,6 @@ class Attribute{
         this.configuration       = {};
         this.fileTypePath        = '';
         this.baseDirFields       = '';
-        this.defaultValues       = {};
         this.options             = {};
         this.type                = '';
         this.isReference         = (options && options.reference) ? true : false;
@@ -20,10 +18,12 @@ class Attribute{
         this.avaliableTypesReference = ['hasMany','belongsToMany','hasOne','belongsToOne']
     }
 
-    loadConfig(){
-        this.configuration = ConfigManager.getConfigAttributes();
-        this.setDefaultValues(this.configuration.defaultValues || {})
+    setConfig(configContent){
+        this.configuration = configContent;
         this.setDictionary(this.configuration.dictionary || {})
+        
+        if(this.configuration.defaultValues)
+            this.setDefaultValues(this.configuration.defaultValues.attribute || {})
     }
 
     setDefaultValues(defaultValues){
@@ -58,23 +58,21 @@ class Attribute{
             .forEach(setValuesInInstance)
     }
 
-    fromJson(json, name){
+    import(contentFile, name){
         this.name = name;
 
-        if(typeof json != "object") json = JSON.parse(json)
-
-        if(!json.type){
+        if(!contentFile.type){
             console.log(chalk.red(`TYPE not found in attribute`))
             throw new Error('Type not found in attribute')
         }
 
-        if(!json.field && this.configuration.field){
+        if(!contentFile.field && this.configuration.field){
             console.log(chalk.yellow(`FIELD not found in attribute (${this.name}). TYPE will be define to working in THIS field`))
         }
 
-        this.type    = json.type;
-        this.field   = json.field;
-        this.options = json;
+        this.type    = contentFile.type;
+        this.field   = contentFile.field;
+        this.options = contentFile;
 
         if(this.options.reference){
             this.getReference(this.options.reference)
@@ -93,19 +91,17 @@ class Attribute{
                 throw new Error('File not found')
             }
         }
-
-        this.loadConfig();
     }
 
     getReference(options){
         if(this.isReference) return;
 
         if(typeof options == 'string'){
-            let optionSlited = options.split('|')
+            let [entity, relationship] = options.split('|')
 
             options              = {}
-            options.entity       = optionSlited[0]
-            options.relationship = optionSlited[1]
+            options.entity       = entity.trim()
+            options.relationship = relationship.trim()
         }
 
 
@@ -126,9 +122,9 @@ class Attribute{
 
         this.field = (this.field) ? this.field : options.relationship
 
-        let referenceEntity = fs.readFileSync(`${this.baseDir}${this.configurationPlugin.entity_dir}/${options.entity}.json`,'utf-8')
+        let referenceEntity = require(`${this.baseDir}${this.configurationPlugin.entity_dir}/${options.entity}.js`)
         this.reference = new EntityManager(this.configurationPlugin, this.baseDir, this.render, { reference:true })
-        this.reference.fromJson(referenceEntity)
+        this.reference.import(referenceEntity)
         this.isReference = true;
         this.typeReference = options.relationship;